Initial commit
Vectorized BEMT function for MATLAB/Octave. Takes a function handle for the lift and drag data of the form [cl cd] = liftfunc(alpha). Alpha should be in radians unless you want to do a conversion in the lift function itself.

Function iterates to calculate induced velocity, thrust, torque, and power (in Watts).